+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# 2.1
+
+2022-09-04
+
+* Added buttons in the left side panels to edit image headers
+    - This may be useful for tweaking things (e.g. bump reference pixel values to shift an HST image alignment), to declutter HISTORY and COMMENT cards, etc.  You can also try deleting offending header cards if something is preventing proper image combination.
+    - Currently, after making edits in the popup window the user must also click the Apply Header Changes button to apply the changes.  Clicking Edit Header again before applying will discard the changes. 
+    - Headers can also be saved to / loaded from text files
+* Fixed the Save Image button functionality 
+* Added options for setting xaxislabel and yaxislabel in the convenience functions plotsinglemulticolorRGB() and comparemulticolorRGB_pureRGB()
+* Added fields for x/y axis ticklabel formats.  (See [the astropy documentation](https://docs.astropy.org/en/stable/visualization/wcsaxes/ticks_labels_grid.html#tick-label-format for options) for available formats and syntax)
+* Combined image x/y axis labels now automatically generate from header CTYPE1/2 values (when present)
+* Added a button for turning minor ticks on/off
+* Added simple gaussian smoothing function.  This can be useful for reducing noise, matching resolutions, etc.
+
+
 # 2.0
 
 2019-09-12